fix compile time warning (basically just added a cast)

+++ src/backend/tcop/pquery.c
@@ -327,7 +327,7 @@ ProcessQueryDesc(queryDesc)
      * ----------------
      */
     if (isRetrieveIntoRelation)
-    	QdSetDest(queryDesc, None);
+    	QdSetDest(queryDesc, (int) None);
     
     /* ----------------
      *	create a default executor state..
